XML Parser 1.0 uploaded
Usage: $ python XML_Parse.py

-Only a few known bugs, if messing with it and find more bugs, please
report them at the top of the file.

-Might be worth it to make a branch when messing with this file, due to
the high number of potential edge-cases in each PDF given the year. If
there is multiple edge-cases throughout the last 10 years, we will need
to compile a list of them and add them individually to the file

-Important: When debugging, printing this file to the console takes up
a lot of space, so it may be hard to find something in the terminal.
You have been warned.

-Should take no more than 20 sec. to run, depending on file size.