Fixed part of the database
See ISSUES for more detailed info…

Basically there is a bug which is currently unfixable due to the lack
of information on the PDF side of things.

The workaround as of now is to simply change the information locally in
the database itself until I can find a workaround.

Thankfully this only affects one dept_short_name in one department, so
it is pretty easy to fix