Map Imaging Tools
In the case where an iframe on hovering over the map links is either
inefficient, or doesn’t look very good, here is a workaround.

There is python file that runs selenium in order to access the
map-links, close the black box, and take a screenshot of the page
(Map-Images 1 and 2).

Then, in a Photoshop Script (not included), I ran all of the images
into a default image crop, so all of the images are zoomed to the same
level, and cropped to be the same size.

There are two images sizes included, and I can create more too, but one
set of images is out 50mb, while the other is about 30mb. If 50mb
doesn’t take too long to load on each page, it is definitely the better
way to go.